changes fake to ok locations
