+++ AirCasting/Map/GoogleMapView.swift
@@ -5,7 +5,6 @@
 //  Created by Lunar on 26/01/2021.
 //
 
-import Foundation
 import UIKit
 import SwiftUI
 import GoogleMaps
